Add utils.sh
- Add utils.sh and move common code to it

The dhall github releases page dl does not
match the original `grep` command because
`os_arch` gets set to `linux` and dhall's
page seems to use a capital 'Linux'.

I made a commit to fix this already, but it
only changed it in one place. Instead of
changing it in 2 places, I decided to fix
the code duplication.